+++ packages/blip/app/components/chart/stats.js
@@ -244,13 +244,11 @@ class Stats extends React.Component {
         isAutomatedBasalDevice && addStat(commonStats.timeInAuto)
         addStat(commonStats.carbs)
         cbgSelected && addStat(commonStats.standardDev)
-        cbgSelected && addStat(commonStats.coefficientOfVariation)
         break
 
       case 'trends':
         addStat(commonStats.averageGlucose)
         addStat(commonStats.standardDev)
-        addStat(commonStats.coefficientOfVariation)
         break
 
       case 'patientStatistics':
@@ -260,7 +258,6 @@ class Stats extends React.Component {
         isAutomatedBasalDevice && addStat(commonStats.timeInAuto)
         addStat(commonStats.carbs)
         cbgSelected && addStat(commonStats.standardDev)
-        cbgSelected && addStat(commonStats.coefficientOfVariation)
         break
     }