+++ test/locales/translations.test.js
@@ -170,12 +170,13 @@ const ignoredTransKeyForYourLoops = [
   'compute-oneday-time-in-range',
   'remaining-time',
   'show-more',
-  'glucose-management-indicator-empty-stat',
   'created-by-unknown',
   'delete-medical-report-number',
   'filters-activated',
   'remote-monitoring',
-  'no-settings-on-device-alert-message'
+  'no-settings-on-device-alert-message',
+  'BGM',
+  'tooltip-empty-stat'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   'yourloops|${s}',